fix(cabi): native 64-bit template marshaling (u32 -> usize) — compliance 28→139
The C ABI marshaled attribute templates as u32 triples [type, ptr, len] and
embed_ptr rejected any non-null pValue on 64-bit (CKR_FUNCTION_FAILED). That
made every value-carrying op fail on native while working in wasm32 (32-bit
pointers fit u32). Widen the template triple element u32 -> usize:

- ck_abi.rs: xlate_template -> Vec<usize> (carry full pValue), tmpl_ptr,
  C_GetAttributeValue length write-back (native CK_UNAVAILABLE_INFORMATION).
- crypto/handlers.rs: get_attr_ulong / get_attr_bytes / absorb_template_attrs
  read usize triples; cast type back to u32.
- ffi.rs: the 8 inline template consumers (CreateObject, FindObjectsInit,
  DeriveKey, CopyObject, SetAttributeValue, GetAttributeValue, …) read usize
  triples; cast type/len to u32, reconstruct pValue at full width.

Portable: usize == u32 on wasm32 (byte-identical, in-browser HSM unchanged),
u64 on native. The validated native:: path is untouched.

Gate (native macOS arm64):
- p11_v32_compliance_test: 28 PASS -> 139 PASS (0/0/0 → 139/66/26).
- 1452 KMIP interop + KAT replay: 15/15 GREEN (zero regression).

Mechanism-param marshaling + missing primitives + conformance tail: follow-ups.

+++ rust/src/crypto/handlers.rs
@@ -59,9 +59,9 @@ pub unsafe fn get_attr_ulong(template: *mut u8, count: u32, attr_type: u32) -> O
     if count > 65536 {
         return None; // Guard against malformed templates with huge count values
     }
-    let ptr = template as *mut u32;
+    let ptr = template as *mut usize;
     for i in 0..count {
-        let t = *ptr.add((i * 3) as usize);
+        let t = *ptr.add((i * 3) as usize) as u32;
         if t == attr_type {
             let val_ptr = *ptr.add((i * 3 + 1) as usize) as usize as *const u32;
             if !val_ptr.is_null() {
@@ -82,9 +82,9 @@ pub unsafe fn get_attr_bytes(template: *mut u8, count: u32, attr_type: u32) -> O
     if count > 65536 {
         return None; // Guard against malformed templates with huge count values
     }
-    let ptr = template as *mut u32;
+    let ptr = template as *mut usize;
     for i in 0..count {
-        let t = *ptr.add((i * 3) as usize);
+        let t = *ptr.add((i * 3) as usize) as u32;
         if t == attr_type {
             let val_ptr = *ptr.add((i * 3 + 1) as usize) as usize as *const u8;
             let val_len = *ptr.add((i * 3 + 2) as usize) as usize;
@@ -143,9 +143,9 @@ pub unsafe fn absorb_template_attrs(attrs: &mut Attributes, template: *mut u8, c
     if template.is_null() || count == 0 || count > 65536 {
         return;
     }
-    let ptr = template as *mut u32;
+    let ptr = template as *mut usize;
     for i in 0..count {
-        let attr_type = *ptr.add((i * 3) as usize);
+        let attr_type = *ptr.add((i * 3) as usize) as u32;
         let val_ptr = *ptr.add((i * 3 + 1) as usize) as usize as *const u8;
         let val_len = *ptr.add((i * 3 + 2) as usize) as usize;
         // Skip key/seed material, internal private attrs, and server-managed attrs.
